requirements.txt
openai-whisper
moviepy == 1.0.3
openai == 1.30.2